requirements.txt
python-dotenv==0.20.0
Pillow==9.5.0
aiohttp==3.8.1
Pyrogram==2.0.106
pytz==2022.1
requests==2.27.1
TgCrypto==1.2.5